requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
duckdb>=0.8.0
python-dotenv>=0.19.0
pydantic>=1.8.2
thefuzz>=0.19.0
starlette>=0.14.2
asyncio>=3.4.3 